Extract consts and set Link header globally
