+++ packages/shopify-app-remix/src/auth/admin/__tests__/session-token-header-path.test.ts
@@ -12,7 +12,7 @@ import {
   signRequestCookie,
   testConfig,
 } from '../../../__tests__/test-helper';
-import {REAUTH_URL_HEADER} from '../../helpers/add-response-headers';
+import {REAUTH_URL_HEADER} from '../../const';
 
 describe('authorize.session token header path', () => {
   describe('errors', () => {
