+++ packages/shopify-app-remix/src/auth/helpers/__tests__/add-response-headers.test.ts
@@ -1,4 +1,5 @@
 import {shopifyApp} from '../../..';
+import {APP_BRIDGE_NEXT_URL} from '../../const';
 import {
   APP_URL,
   TEST_SHOP,
