+++ packages/shopify-app-remix/src/billing/__tests__/cancel.test.ts
@@ -20,7 +20,7 @@ import {
   testConfig,
 } from '../../__tests__/test-helper';
 import {mockExternalRequest} from '../../__tests__/request-mock';
-import {REAUTH_URL_HEADER} from '../../auth/helpers/add-response-headers';
+import {REAUTH_URL_HEADER} from '../../auth/const';
 
 import * as responses from './mock-responses';
