+++ packages/shopify-app-remix/src/billing/__tests__/request.test.ts
@@ -24,7 +24,7 @@ import {
   mockExternalRequest,
   mockExternalRequests,
 } from '../../__tests__/request-mock';
-import {REAUTH_URL_HEADER} from '../../auth/helpers/add-response-headers';
+import {REAUTH_URL_HEADER} from '../../auth/const';
 
 import * as responses from './mock-responses';
